+++ package.json
@@ -11,6 +11,7 @@
     "react": "^16.14.0",
     "react-dom": "^16.14.0",
     "react-instagram-embed": "^2.0.0",
+    "react-router-dom": "^5.2.0",
     "react-scripts": "4.0.3"
   },
   "scripts": {
